Add script to update appdata column in cel table.
For the update to works on every xivo, postgres must be restarted just
before the "alter table" is run because there is a case where a pending
transaction holds a lock on the cel table, making the upgrade hangs.
This happens if there was an error in the cel_pgsql asterisk module.