remove old FAI package before the actual upgrade but after getting upgrade confirmation
Problem was: Keeping the old FAI packages during the upgrade makes
"apt-get upgrade" upgrade old packages (i.e. packages renamed from
pf-xivo-* to xivo-*), resulting in packages being upgraded to their
last version. This results in a mix of different packages being
installed in different versions, and some of them are incompatible.

In this case, the conflict is between pf-xivo-base-config and
xivo-libsccp, because, with the wrong version combination, they both
own /etc/asterisk/sccp.conf.

This results in the non-upgrade of pf-xivo-base-config, blocking the
upgrade of xivo-libsccp which would have been done later by the same
command.

The consequence is that xivo-config can not install, because
xivo-libsccp still owns sccp.conf (because it was not upgraded).

The consequence of that is that packages depending on xivo-config can
not be installed, which is virtually all packages of xivo.