xivo-upgrade: add another test for broken grub
The device.map can be fine and you might still run into trouble after
the upgrade from grub-pc (wheezy to jessie version). So add a test that
detects another case, which is linked to the fact that we hide question
(DEBIAN_FRONTEND=noninteractive) during the upgrade.

There's no way to correctly preseed this question in XiVO.