jessie: handle local fail2ban modification
If /etc/fail2ban/filter.d/asterisk.conf was modified before the upgrade,
this was causing the upgrade to fail, since dpkg though that the file
was still part of the xivo-config package before upgrading the fail2ban
package, which now ships this file in jessie.

One way to make dpkg forget that the file is not part of the xivo-config
package anymore is to reinstall the package.

Another solution would have been to do a "rm_conffile" (instead of a
mv_conffile) in the xivo-config package.