don't force webi installation on upgrade
The goal is that, eventually, the web interface will be completely
optional on XiVO. Right now, if you uninstall the xivo and
xivo-web-interface package (leaving only the xivo-base package and all
its dependencies), on upgrade the xivo and xivo-web-interface package
will be reinstalled.

With this modification, these two packages won't be reinstalled, but
they will still be upgraded if they were installed.

I suspect that the forced "apt-get install xivo" was useful mostly when
we were renaming packages: this way this was forcing apt-get not to
uninstall the xivo package. I believe this is now unnecessary.

Also added a condition on the post-start.d scripts that had a dependency
on the web-interface. This was not strictly necessary, but I feel like
it's clearer.