move pg 9.1 to 9.4 cluster upgrade out
Moved into a separate package, currently named xivo-dbms, that is
responsible to make the postgres cluster upgrade, so that the postgres
server listening on port 5432 is always the "right" one.

Does not move pg 9.0 to 9.1 cluster upgrade since this upgrade is done
during the Debian 6 to Debian 7 upgrade, not on Debian 8 or later.

This is the chosen solution to fix #6274.