purge postgresql-X.X packages in remove_unused_packages
The postgresql-X.X packages are never autoremoved automatically, even if
they were automatically installed (which is not always the case with
xivo, depending if the system has been upgraded or not) because of the
configuration of apt in /etc/apt/apt.conf.d/01autoremove-postgresql.

So these packages need to be manually removed (to keep things clean),
and this can't be done in a postinst since you is not "reentrant", so
this can't be done in xivo-dbms AFAIK. So do it in xivo-upgrade.

Maybe there's something to do with the xivo-dbms package control fields
Provides / Conflicts / Replaces to force the remove of postgresql-X.X
packages, but this seems a bit like playing with fire. And the package
would only be removed and not purged anyway.

When it will be time to purge postgresql-9.4, we'll first have to do the
cluster upgrade in xivo-dbms, and then add a check in
remove_unused_packages that "postgresql-9.4 is installed AND xivo-dbms
version is greater or equal than the version of xivo-dbms that did the
9.4 to 9.X migration" before purging.