add condition to rebuild_provd_config_label_null
When upgrading a XiVO that was installed without the xivo-web-interface
package, the script would produce some gibberish when run. The
script does not be run anyway in recent version of XiVO.

A similar condition was added (commit 935cdb4) to the other scripts that
had a dependency on the web interface, so I feel like this one was
simply forgotten.