add xivo-to-wazo-upgrade
Note that this script is currently not part of the Debian package, since
it does not need to be, but I'm still putting it in the xivo-upgrade
repository since it is somewhat related to use-wazo-infrastructure,
which belong to the xivo-upgrade repository and debian package.

Otherwise, that script is just made to be manually downloaded and
executed.