xivo-to-wazo-upgrade: handle upgrade from archive
Previously, if you were on an archive version (i.e. either with a
"xivo-dist xivo-16.13" for example or a "apt-get install
xivo-fai-14.17"), then when running the xivo-to-wazo-upgrade script, you
weren't proposed a Wazo version. You had to manually switch to a
non-archive version before running the script for it to work properly.

That said, the script is not trying anymore to keep the version
"flavour", i.e it unconditionally switch to the phoenix distribution
(well, you can use the environment variable XIVO_TO_WAZO_DEB_LINE to
bypass it) instead of trying to switch from xivo-dev to wazo-dev and
xivo-rc to wazo-rc. The environment variable is also useful if you want
to upgrade to a Wazo archive version (although a bit verbose, it's quite
flexible).