post-start: restart services after tenant migration
reason: All tokens from backend xivo_service need to be regenerated:
after the post-start scripts, tokens generated when the services
started, i.e. before the post-start scripts, are not associated with any
tenants, but they will after a restart. In Wazo 18.04, a service token
that is associated with no tenants cannot see any user in confd.