force installation of postgresql-9.4 from phoenix dist
reason: We do not want to support upstream version (security) of
postgresql and only freeze version to the phoenix distribution.

Old installation have meta packet 'postgresql' and during the 'apt-get
upgrade' this packet download a superior version than the frozen version
in 'xivo-dbms'. So, the packet manager doesn't known of to solve
conflict when it's time to install xivo-dbms, it doesn't want to
downgrade version ...