dird config: use http api to create default dird configs
the old version published messages to the bus but this was not reliable because the bus connection was to
slow to connect and the messages were missed.

this new version uses the HTTP API and will be executed even if the old script has been executed
if the personal source already exist for a tenant it quit.

the office365 config is only generated if the plugin is installed and enabled